+++ src/nyctea/validators/__init__.py
@@ -29,8 +29,11 @@ Example:
 """
 
 from nyctea.validators.base import Validator, ValidatorMetadata
+from nyctea.validators.builtins.register import register_builtins
 from nyctea.validators.column import ColumnCheck, ColumnParser, ColumnValidator
+from nyctea.validators.decorators import ValidatorDecorator
 from nyctea.validators.frame import FrameCheck, FrameParser, FrameValidator
+from nyctea.validators.registry import Registry, ValidatorRegistry
 
 __all__ = [
     "ColumnCheck",
@@ -39,6 +42,10 @@ __all__ = [
     "FrameCheck",
     "FrameParser",
     "FrameValidator",
+    "Registry",
     "Validator",
+    "ValidatorDecorator",
     "ValidatorMetadata",
+    "ValidatorRegistry",
+    "register_builtins",
 ]